Intro description, 6 rooms, 2 things
I created the introductory paragraph to the game, detailing enough to
give the situation of the character. In addition, I created the rooms
that the player will be navigating in through the first half of the
game. I’ve given him two of his necessary items as well that he will
use throughout the game.