Cheat codes, arcade machine, new actions
I’ve created two actions that interact specifically to the arcade
machine. For this puzzle, in order to win the arcade machine (which is
a container), the player must have the cheat codes (the key) with them.
If they player has the cheat codes, the arcade machine will unlock.
However, it is not saying what it is supposed to say after the action
of “playing with” is complete.