Second Puzzle
I have my second puzzle, but it is not fully complete. The player can
access the cheat codes once the power is on and the arcade machine is
locked without them. However, the arcade machine is not responding to
the command. It is not denying access nor is it permitting.